Add test case for type checking
