+++ pyproject.toml
@@ -11,6 +11,8 @@ python = "^3.10"
 
 [tool.poetry.group.dev.dependencies]
 mypy = "^0.982"
+isort = "^5.10.1"
+pytest = "^7.2.0"
 
 [build-system]
 requires = ["poetry-core"]
